Move FineTuneHeading to separate file
